Move knexfile and db files into src/ folder
