+++ package.json
@@ -16,12 +16,12 @@
     "build": "tsc",
     "build:watch": "tsc && (tsc -w --preserveWatchOutput & nodemon)",
     "dev": "npm run build:watch",
-    "start": "probot run ./lib/src/index.js",
+    "start": "probot run ./lib/index.js",
     "lint": "eslint src/**/*.ts",
     "lint-fix": "eslint src/**/*.ts --fix",
     "check-format": "prettier src/**/*.ts -l",
     "format": "prettier src/**/*.ts --write",
-    "migrations": "knex migrate:latest",
+    "migrations": "knex migrate:latest --knexfile=src/db/knexfile.ts",
     "test": "jest",
     "test:watch": "jest --watch --notify --notifyMode=change --coverage"
   },
