+++ src/db/index.ts
@@ -1,5 +1,5 @@
 import knex from "knex";
-import knexfile from "../knexfile";
+import knexfile from "./knexfile";
 
 const env = process.env.NODE_ENV ?? "development";
 const db = knex(knexfile[env]);
